pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*", "audio_client*"]

[project]
name = "realtime-transcription"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "amazon-transcribe",
  "fastapi",
  "numpy",
  "python-dotenv",
  "python-multipart",
  "sounddevice",
  "uvicorn[standard]",
  "websockets",
]

[dependency-groups]
dev = [
  "bump-my-version",
  "git-cliff",
  "httpx",
  "pyright",
  "pytest",
  "pytest-asyncio",
  "ruff",
]

# ---------------------------------------------------------------------------
# bumpversion
# ---------------------------------------------------------------------------
[tool.bumpversion]
current_version = "0.1.0"
commit = true
tag = true
tag_name = "v{new_version}"

[[tool.bumpversion.files]]
filename = "app/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'current_version = "{current_version}"'
replace = 'current_version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

# ---------------------------------------------------------------------------
# git-cliff
# ---------------------------------------------------------------------------
[tool.git-cliff.remote.github]
owner = "bilardi"
repo = "realtime-transcription"

[tool.git-cliff.changelog]
header = "# Changelog\n\nAll notable changes to this project will be documented in this file.\n\nThe format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).\n\n"
body = """
{% if version %}\
## [{{ version | trim_start_matches(pat="v") }}] - {{ timestamp | date(format="%Y-%m-%d") }}
{% else %}\
## [Unreleased]
{% endif %}\
{% for group, commits in commits | group_by(attribute="group") %}\
### {{ group }}
{% for commit in commits %}\
- {{ commit.message | upper_first }}
{% endfor %}
{% endfor %}\
"""
footer = """
{% for release in releases %}\
{% if release.version %}\
{% if release.previous is defined and release.previous.version is defined %}\
[{{ release.version | trim_start_matches(pat="v") }}]: https://github.com/{{ remote.github.owner }}/{{ remote.github.repo }}/compare/{{ release.previous.version }}...{{ release.version }}
{% else %}\
[{{ release.version | trim_start_matches(pat="v") }}]: https://github.com/{{ remote.github.owner }}/{{ remote.github.repo }}/releases/tag/{{ release.version }}
{% endif %}\
{% else %}\
{% if release.previous is defined and release.previous.version is defined %}\
[Unreleased]: https://github.com/{{ remote.github.owner }}/{{ remote.github.repo }}/compare/{{ release.previous.version }}...HEAD
{% endif %}\
{% endif %}\
{% endfor %}
"""
trim = true
render_always = false
skip_if_empty = true

[tool.git-cliff.git]
conventional_commits = true
filter_unconventional = true
tag_pattern = "v[0-9].*"
sort_commits = "oldest"
split_commits = false

[[tool.git-cliff.git.commit_parsers]]
message = "^feat"
group = "Features"

[[tool.git-cliff.git.commit_parsers]]
message = "^fix"
group = "Bug Fixes"

[[tool.git-cliff.git.commit_parsers]]
message = "^refactor|^perf"
group = "Performance"

[[tool.git-cliff.git.commit_parsers]]
message = "^docs"
group = "Documentation"

[[tool.git-cliff.git.commit_parsers]]
message = "^test"
group = "Test"

[[tool.git-cliff.git.commit_parsers]]
message = "^chore"
group = "Chores"

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

# ---------------------------------------------------------------------------
# ruff
# ---------------------------------------------------------------------------
[tool.ruff]
exclude = [".venv"]
include = ["scripts/*.py", "app", "audio_client"]
fix = true
line-length = 100
output-format = "grouped"
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "COM812", # missing trailing comma # conflict with formatter
  "D203", # incorrect-blank-line-before-class
  "D213", # multi-line-summary-second-line
  "D413", # Missing blank line after last section ("Args", "Returns")
  "D415", # First line should end with a period, question mark, or exclamation point
  "EXE002", # The file is executable but no shebang is present
  "FBT", # positional argument in function definition or call
  "ISC001", # implicit string concatenation # conflict with formatter
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["D104"] # Missing docstring in public package
"tests/**" = [
  "ANN", # Missing return / type annotation for public function
  "S101" # Use of `assert` detected
]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 100
indent-style = "space"
line-ending = "auto"
quote-style = "double"

[tool.ruff.lint.mccabe]
max-complexity = 10

# ---------------------------------------------------------------------------
# pyright
# ---------------------------------------------------------------------------
[tool.pyright]
exclude = ["**/__pycache__"]
include = ["scripts", "app", "audio_client"]
pythonVersion = "3.13"
reportMissingTypeStubs = false
typeCheckingMode = "strict"
